config.rsのモジュール分離
